+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Remove `idoitapi.php` because Composer is the prefered way to use
 -   Require entry identifier in methods `CMDBCategory::archive()`, `delete()` and `purge()`
 -   Methods `cmdb.category.create`, `cmdb.category_info.read` (and others, too) do not need parameters `catg` or `cats`. Parameter `category` seems to be sufficient.
+-   Make `CMDBCategory::purge()` a lot faster due to method `cmdb.category.quickpurge`
 -   Return empty array for reports with no results (class `CMDBReports`)
 -   Remove many dependencies from unit tests